folder restructure
- goal of mentally providing boundries between parts of the app that should interface but not be dependent on the implementation of other parts
