+++ esbuild.config.mjs
@@ -14,7 +14,7 @@ const context = await esbuild.context({
 	banner: {
 		js: banner,
 	},
-	entryPoints: ["src/main.ts"],
+	entryPoints: ["src/obsidian/main.ts"],
 	bundle: true,
 	external: [
 		"obsidian",
