--- src/select-vault/index.tsx
@@ -1,1 +1,0 @@
-export * from "./select-vault-modal";
